# app/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FEM 巡檢管理系統 API",
    "description": "設施設備巡檢管理系統的 RESTful API 文件",
    "version": "1.0.0",
    "contact": {
      "name": "API 支援",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API 伺服器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "輸入您的 JWT Token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/inspection/statistics": {
      "get": {
        "tags": ["巡檢查詢"],
        "summary": "取得巡檢統計資料",
        "description": "取得儀表板統計資訊",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "查詢日期（預設為今日）",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得統計資料",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "abnormal_tracking": {
                          "type": "object"
                        },
                        "inspection_tasks": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/inspection/records": {
      "get": {
        "tags": ["巡檢查詢"],
        "summary": "查詢巡檢記錄",
        "description": "查詢歷史巡檢記錄",
        "parameters": [
          {
            "name": "org_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "route_id",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "inspector_id",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "result_status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["正常", "異常", "未檢查"]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功取得巡檢記錄",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "records": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "pagination": {
                          "type": "object"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}